refactor!: extract the Tree component into lattice-php/tree
The Tree view component moves to its own first-party package,
lattice-php/tree, so it can grow lazy child loading and drag & drop on
its own release cadence without expanding the core surface.

Removed from core: the Tree component, TreeNode, the TreeSource contract,
CallbackTreeSource and EloquentTreeSource, the React renderer and its
context, all Tree tests and the docs page/fixture, the workbench demo
page/action/menu entry, and the `common.tree.*` strings (they now live in
the package's own `tree` i18next namespace). The `ScopedEloquentQuery`
trait had only these two consumers, so it is inlined back into its
remaining user, EloquentOptions.

BREAKING CHANGE: The Tree component is no longer part of lattice-php/lattice.
Run `composer require lattice-php/tree` and the discovered plugin registers
it automatically; the `Tree`, `TreeNode`, and source classes move from the
`Lattice\Lattice\Ui` namespace to the flat `Lattice\Tree` namespace.
