+++ docs/content/docs/components/overview.mdx
@@ -28,19 +28,18 @@ serialized **Tree** the renderer receives, and the theme **Style** tokens it use
 
 ## Catalog
 
-| Component                                      | Purpose                                                               | Page                                                      |
-| ---------------------------------------------- | --------------------------------------------------------------------- | --------------------------------------------------------- |
-| `Stack`, `Grid`, `Card`                        | Arrange children — vertical/horizontal, columns, bordered panels      | [Layout](/components/layout/)                             |
-| `Section`, `Collapsible`                       | Titled and untitled disclosures that fold open and closed             | [Section & Collapsible](/components/section-collapsible/) |
-| `FloatingPanel`                                | Pin children to a viewport corner                                     | [Floating panel](/components/floating-panel/)             |
-| `Heading`, `Text`, `Badge`, `Icon`, `RawBlock` | Display primitives — headings, prose, pills, sprite icons, raw HTML   | [Text & badges](/components/text/)                        |
-| `Button`, `Link`, `SegmentedControl`           | Buttons, navigational links, and standalone segmented controls        | [Buttons & links](/components/buttons/)                   |
-| `Tabs`, `Tab`                                  | A tab strip with URL-synced, lazily rendered panels                   | [Tabs](/components/tabs/)                                 |
-| `Modal`                                        | A dialog opened and closed by action effects                          | [Modals](/components/modals/)                             |
-| `Tooltip`                                      | An ⓘ info popover you can place anywhere                              | [Tooltip](/components/tooltip/)                           |
-| `Chart`                                        | Line, bar, area, and pie charts                                       | [Charts](/components/charts/)                             |
-| `Tree`                                         | A keyboard-navigable hierarchy with icons, badges, links, and actions | [Tree](/components/tree/)                                 |
-| `Notifications`                                | A bell listing a user's database notifications, realtime + polling    | [Notifications](/components/notifications/)               |
+| Component                                      | Purpose                                                             | Page                                                      |
+| ---------------------------------------------- | ------------------------------------------------------------------- | --------------------------------------------------------- |
+| `Stack`, `Grid`, `Card`                        | Arrange children — vertical/horizontal, columns, bordered panels    | [Layout](/components/layout/)                             |
+| `Section`, `Collapsible`                       | Titled and untitled disclosures that fold open and closed           | [Section & Collapsible](/components/section-collapsible/) |
+| `FloatingPanel`                                | Pin children to a viewport corner                                   | [Floating panel](/components/floating-panel/)             |
+| `Heading`, `Text`, `Badge`, `Icon`, `RawBlock` | Display primitives — headings, prose, pills, sprite icons, raw HTML | [Text & badges](/components/text/)                        |
+| `Button`, `Link`, `SegmentedControl`           | Buttons, navigational links, and standalone segmented controls      | [Buttons & links](/components/buttons/)                   |
+| `Tabs`, `Tab`                                  | A tab strip with URL-synced, lazily rendered panels                 | [Tabs](/components/tabs/)                                 |
+| `Modal`                                        | A dialog opened and closed by action effects                        | [Modals](/components/modals/)                             |
+| `Tooltip`                                      | An ⓘ info popover you can place anywhere                            | [Tooltip](/components/tooltip/)                           |
+| `Chart`                                        | Line, bar, area, and pie charts                                     | [Charts](/components/charts/)                             |
+| `Notifications`                                | A bell listing a user's database notifications, realtime + polling  | [Notifications](/components/notifications/)               |
 
 ## Where components appear
