--- tests/Feature/Core/Components/TreeTest.php
@@ -1,113 +1,0 @@
-<?php
-declare(strict_types=1);
-
-use Lattice\Lattice\Support\Wire;
-use Lattice\Lattice\Ui\Components\Tree;
-use Lattice\Lattice\Ui\Sources\CallbackTreeSource;
-use Lattice\Lattice\Ui\Values\TreeNode;
-
-it('serializes an eager node tree with defaults', function (): void {
-    $node = wire(
-        Tree::make()->nodes([
-            TreeNode::make('Electronics', '1')->children([TreeNode::make('Laptops', '2')]),
-        ]),
-    );
-
-    expect($node['type'])->toBe('tree')
-        ->and($node['props']['nodes'][0])->toMatchArray(['id' => '1', 'label' => 'Electronics'])
-        ->and($node['props']['nodes'][0]['children'][0])->toMatchArray(['id' => '2', 'label' => 'Laptops'])
-        ->and($node['props']['rememberState'])->toBeFalse()
-        ->and($node['props']['defaultExpanded'])->toBe([]);
-});
-
-it('serializes activeId, defaultExpanded, and rememberState', function (): void {
-    $node = wire(
-        Tree::make()->nodes([TreeNode::make('A', '1')])->activeId('1')->defaultExpanded(['1'])->rememberState(),
-    );
-
-    expect($node['props'])->toMatchArray([
-        'activeId' => '1', 'defaultExpanded' => ['1'], 'rememberState' => true,
-    ]);
-});
-
-it('serializes source children recursively for hasChildren nodes', function (): void {
-    $childrenByParent = [
-        'root' => [TreeNode::make('Child', 'child')->hasChildren()],
-        'child' => [TreeNode::make('Grandchild', 'grandchild')],
-    ];
-
-    $node = wire(
-        Tree::make()->source(new CallbackTreeSource(
-            roots: fn (): array => [TreeNode::make('Root', 'root')->hasChildren()],
-            children: fn (string $parentId): array => $childrenByParent[$parentId] ?? [],
-        )),
-    );
-
-    $root = $node['props']['nodes'][0];
-    expect($root)->toMatchArray(['id' => 'root', 'hasChildren' => true])
-        ->and($root['children'][0])->toMatchArray(['id' => 'child'])
-        ->and($root['children'][0]['children'][0])->toMatchArray(['id' => 'grandchild', 'label' => 'Grandchild']);
-});
-
-it('stops fetching source children at the depth cap so cyclic data terminates', function (): void {
-    $fetches = 0;
-
-    $node = wire(
-        Tree::make()->source(new CallbackTreeSource(
-            roots: fn (): array => [TreeNode::make('Root', 'n0')->hasChildren()],
-            children: function (string $parentId) use (&$fetches): array {
-                $fetches++;
-
-                return [TreeNode::make('Child', "n{$fetches}")->hasChildren()];
-            },
-        )),
-    );
-
-    $boundary = $node['props']['nodes'][0];
-    while (isset($boundary['children'])) {
-        $boundary = $boundary['children'][0];
-    }
-
-    expect($fetches)->toBe(50)
-        ->and($boundary)->toMatchArray(['id' => 'n50', 'hasChildren' => true]);
-});
-
-it('truncates eager children at the depth cap with a hasChildren marker', function (): void {
-    $subtree = TreeNode::make('Leaf', 'leaf');
-    foreach (range(51, 0) as $level) {
-        $subtree = TreeNode::make("Level {$level}", "n{$level}")->children([$subtree]);
-    }
-
-    $node = wire(Tree::make()->nodes([$subtree]));
-
-    $boundary = $node['props']['nodes'][0];
-    while (isset($boundary['children'])) {
-        $boundary = $boundary['children'][0];
-    }
-
-    expect($boundary)->toMatchArray(['id' => 'n50', 'hasChildren' => true]);
-});
-
-describe('docs fixtures', function (): void {
-    it('matches the tree example fixture', function (): void {
-        assertFixtureMatches('components.tree', sortFixtureKeys(stripFixtureRefs(Wire::toWire([
-            Tree::make('category-tree')
-                ->nodes([
-                    TreeNode::make('Electronics', 'electronics')
-                        ->icon('cpu')
-                        ->children([
-                            TreeNode::make('Laptops', 'electronics-laptops'),
-                            TreeNode::make('Phones', 'electronics-phones')->href('/products/phones'),
-                        ]),
-                    TreeNode::make('Clothing', 'clothing')
-                        ->badge('New')
-                        ->children([
-                            TreeNode::make('Men', 'clothing-men'),
-                            TreeNode::make('Women', 'clothing-women'),
-                        ]),
-                ])
-                ->activeId('electronics-phones')
-                ->defaultExpanded(['electronics']),
-        ]))));
-    });
-});
