--- workbench/app/Actions/ShowTreeNodeInfoAction.php
@@ -1,26 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Workbench\App\Actions;
-
-use Illuminate\Http\Request;
-use Lattice\Lattice\Actions\ActionDefinition;
-use Lattice\Lattice\Actions\ActionResult;
-use Lattice\Lattice\Actions\Components\Action as ActionComponent;
-use Lattice\Lattice\Attributes\AsAction;
-
-#[AsAction('workbench.tree.show-node-info')]
-class ShowTreeNodeInfoAction extends ActionDefinition
-{
-    public function definition(ActionComponent $action): ActionComponent
-    {
-        return $action
-            ->label(__('workbench.actions.tree-node-info.label'))
-            ->icon('info');
-    }
-
-    public function handle(Request $request): ActionResult
-    {
-        return ActionResult::success()->openModal('tree-node-info');
-    }
-}
