+++ workbench/lang/en/workbench.php
@@ -185,9 +185,6 @@ return [
             'label' => 'Reject selected',
             'toast' => 'Rejected :count products: :reason',
         ],
-        'tree-node-info' => [
-            'label' => 'About this tree',
-        ],
     ],
     'auth' => [
         'login' => [
@@ -421,7 +418,6 @@ return [
         'tables' => 'Tables',
         'tabs' => 'Tabs',
         'toggle-sidebar' => 'Toggle sidebar',
-        'tree' => 'Tree',
     ],
     'pages' => [
         'charts' => [
@@ -568,80 +564,6 @@ return [
                 'bars' => 'Progress bars',
                 'circles' => 'Progress circles',
                 'title' => 'Progress',
-            ],
-            'tree' => [
-                'categories' => [
-                    'automotive' => [
-                        'accessories' => [
-                            'floor-mats' => 'Floor mats',
-                            'label' => 'Accessories',
-                            'seat-covers' => 'Seat covers',
-                        ],
-                        'label' => 'Automotive',
-                        'parts' => [
-                            'brakes' => 'Brakes',
-                            'engine' => 'Engine',
-                            'label' => 'Parts',
-                        ],
-                    ],
-                    'clothing' => [
-                        'accessories' => [
-                            'belts' => 'Belts',
-                            'hats' => 'Hats',
-                            'label' => 'Accessories',
-                        ],
-                        'kids' => 'Kids',
-                        'label' => 'Clothing',
-                        'men' => 'Men',
-                        'women' => 'Women',
-                    ],
-                    'documents' => [
-                        'archive' => 'Archive',
-                        'label' => 'Documents',
-                        'rename' => 'Rename',
-                    ],
-                    'electronics' => [
-                        'accessories' => [
-                            'cases' => 'Cases',
-                            'chargers' => 'Chargers',
-                            'label' => 'Accessories',
-                        ],
-                        'label' => 'Electronics',
-                        'laptops' => 'Laptops',
-                        'phones' => 'Phones',
-                        'tablets' => 'Tablets',
-                    ],
-                    'furniture' => [
-                        'bedroom' => [
-                            'beds' => 'Beds',
-                            'label' => 'Bedroom',
-                            'wardrobes' => 'Wardrobes',
-                        ],
-                        'label' => 'Furniture',
-                        'living-room' => [
-                            'coffee-tables' => 'Coffee tables',
-                            'label' => 'Living room',
-                            'sofas' => 'Sofas',
-                        ],
-                    ],
-                    'groceries' => [
-                        'bakery' => 'Bakery',
-                        'dairy' => 'Dairy',
-                        'frozen' => 'Frozen',
-                        'label' => 'Groceries',
-                        'produce' => 'Produce',
-                    ],
-                    'help' => [
-                        'label' => 'Help',
-                    ],
-                ],
-                'info-modal' => [
-                    'body' => 'Arrow keys move focus between rows without leaving the page tab order. Enter or Space activates the focused row: it follows a link, runs an action, or marks the row active.',
-                    'description' => 'How keyboard navigation works in this component.',
-                    'title' => 'Keyboard navigation',
-                ],
-                'intro' => 'A keyboard-navigable hierarchy with active state, a linked node, row actions, and a modal-opening action.',
-                'title' => 'Tree',
             ],
         ],
         'dependent' => [
